pyproject.toml
[tool.uv.sources]
torch = [
  { index = "pytorch-cu129", marker = "sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu129", marker = "sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[project]
name = "contract-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.53.0",
    "ipykernel>=6.30.1",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.0.0",
    "supervision>=0.26.1",
    "timm>=1.0.20",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.2",
    "ultralytics>=8.3.202",
]
